requirements.txt
asgiref==3.8.1
av==12.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.4
django-allauth==0.63.6
djangorestframework==3.15.1
future==1.0.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
proglog==0.1.10
pyee==11.1.0
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
